requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/damnever/pigar

# bot.py: 8
APScheduler == 3.5.1

# bot.py: 2
# picture_pallet.py: 1
# test.py: 1
Pillow == 3.1.2

# picture_pallet.py: 3
numba == 0.38.1

# bot.py: 7
python_firebase == 1.2

# bot.py: 1
python_twitter == 3.3
